% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_init.R
\name{pr_closeConnection}
\alias{pr_closeConnection}
\title{Closes database connection}
\usage{
pr_closeConnection()
}
\value{
None

This function closes an existing database connection
}
\description{
This function closes the current database connection
}
