% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_createTables.R
\name{pr_createTables}
\alias{pr_createTables}
\title{Creates the psqlrepo database tables}
\usage{
pr_createTables(delete = FALSE)
}
\arguments{
\item{delete}{if true, existing tables are deleted}
}
\description{
This function creates the necesarry psqlrepo database tables within
the current psqlrepo schema. It returns an error if the schema does not exists
}
