% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_deleteKey.R
\name{pr_deleteKey}
\alias{pr_deleteKey}
\title{Delete meta data key for an object}
\usage{
pr_deleteKey(name, key)
}
\arguments{
\item{name}{name of object}

\item{key}{name of key

This function deletes a key for a given object. Both key and value
are removed from the database.}
}
\value{
TRUE if successfull, FALSE otherwise
}
\description{
Deletes a meta data key, and value, for an given object
}
