% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_deleteKeyObj.R
\name{pr_deleteKeyObj}
\alias{pr_deleteKeyObj}
\title{Delete all keys from objects}
\usage{
pr_deleteKeyObj(name)
}
\arguments{
\item{name}{name of object

This function deletes al meta data from an
object. All keys and associated values are removed from the
database. Be carefull with this function since the deletion is
permanent and there is now way back}
}
\description{
This function deletes all meta data (keys and values) from an
object
}
