% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_deleteObj.R
\name{pr_deleteObj}
\alias{pr_deleteObj}
\title{Delete a regular object form the repository}
\usage{
pr_deleteObj(name)
}
\arguments{
\item{name}{of te object to delete}
}
\value{
NULL if succesfull

This function deletes stuf, but needs some more explaining in the
man-page
}
\description{
This functions deletes a regular object from the repository,
including meta data.
}
