% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_createTables.R
\name{pr_dropTables}
\alias{pr_dropTables}
\title{Drop the psqlrepo database tables}
\usage{
pr_dropTables()
}
\description{
This function drops the psqlrepo database tables within
the current psqlrepo schema. It returns an error if the schema or
tables donot exists
}
