% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_getConfig.R
\name{pr_getConfig}
\alias{pr_getConfig}
\title{Get psqlrepo configuration}
\usage{
pr_getConfig()
}
\value{
A list with configuration options

This function is used from other functions of this package. It is
not intended to use directly. However, it can be useful in some
cases like debugging of the configuration.
}
\description{
The configuration options are stored in a separate environment.
This function retrieves these options
}
