% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_getKey.R
\name{pr_getKey}
\alias{pr_getKey}
\title{Get meta data key and values from object in  repository}
\usage{
pr_getKey(name, key)
}
\arguments{
\item{name}{Name of the object.}

\item{key}{Key.}
}
\value{
A character object with the value of the key, 

Meta data is stored as key, value pairs. This function associates a
key, value pair with the given object (objectname). Each object
can have an unlimited number of keys and it is up to the user to
organise the keys and possible values,

Values are always stored as character strings, so numeric values
must by changed into strings before they can be stored.

This function retrieves the value for an existing key from an
object stored in the repository. If the object or key does not
exists, it returns an error.
}
\description{
Meta data is stored as key,value pairs in the repository, This
function retrieves a key for a specific object
}
