% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_getKeyObj.R
\name{pr_getKeyObj}
\alias{pr_getKeyObj}
\title{Get meta data keys and values from specific object}
\usage{
pr_getKeyObj(name)
}
\arguments{
\item{name}{name of object}
}
\value{
a list with key-value pairs

This function retrieves all the keys for an object and the
associated values. This information is returned as a list. The
names of this list are equal to the keys, the values are stored as
list values for each key.
}
\description{
This function retrieves the mete data keys from an object in the
repository
}
