% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_getObj.R
\name{pr_getObj}
\alias{pr_getObj}
\title{Get an object from the repository}
\usage{
pr_getObj(name, verbose = FALSE, ignoretype = FALSE)
}
\arguments{
\item{name}{store object using thisname}

\item{verbose}{give some verbose output}

\item{ignoretype}{ignore the object type}
}
\value{
The R object as stored in the repository

This function gets an regular object from the database. The object
is returned and is equal as the object when it was stored
So if you store a list in the repository,
this function wil return that list.

In the repository different types of objects can be stored, like
regular objects, blobs or audit objects. This function checks the
type of the object and returns an error if the object is not of a
regular type. When ingnoretype=FALSE this type checking is skipped.
This is useful if you want to inspect an object for e.g. debugging.
}
\description{
This functions gets an regular object from the repository and
returns it
}
