% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_init.R
\name{pr_init}
\alias{pr_init}
\title{Initialise psql repository}
\usage{
pr_init(configfile = "~/.psqlrepo.yml", configuration = "default")
}
\arguments{
\item{configfile}{filename of the configuration file}

\item{config}{name of configuration to read}
}
\value{
TBD

This function also creates a database connection and stores the
handle in the package namespace. Only one connection is possible.
If pr_init is called while a database connection exists, the
current connection is closed and a new connection is created.
}
\description{
This function initialises the psql repository. It loads the
configuration, connects to the database and create a DBI connection
object.
}
