% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_keyExists.R
\name{pr_keyExists}
\alias{pr_keyExists}
\title{Check if a key exists for specific object.}
\usage{
pr_keyExists(name, key)
}
\arguments{
\item{name}{Name of the object.}

\item{key}{Key.}
}
\value{
TRUE if key exists, FALSE otherwise
}
\description{
This function checks if a key exists for a specific object.
}
