% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_objectExists.R
\name{pr_objectExists}
\alias{pr_objectExists}
\title{Check if object exists in database}
\usage{
pr_objectExists(name)
}
\arguments{
\item{objectname}{name of object}
}
\value{
TRUE if object exists, FALSE otherwise
}
\description{
This functions  checks if an object exists in the repository, if
so, it returns TRUE.
}
