% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_readcfg.R
\name{pr_readcfg}
\alias{pr_readcfg}
\title{Read psqlrepo configuration file}
\usage{
pr_readcfg(filename, configuration)
}
\arguments{
\item{filename}{filename of the configuration file}

\item{config}{name of configuration to read}
}
\value{
A configuration object, a list with the configuration options

This function is based on the config::get command from the config
package. While config::get is making assumptions about where to
find the configuration file and which configuration to read, the
pr_readcfg function requires that both filename and
configuration are given explicitly.
}
\description{
The psqlrepo configuration file is a YAML file with database
connection info. This function reads this file and returns a
configuration object.
}
