% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_sql.R
\name{pr_sql}
\alias{pr_sql}
\title{SQL wrapper function}
\usage{
pr_sql(query, verbose = FALSE, errors = TRUE)
}
\arguments{
\item{query}{The SQL query}

\item{verbose}{If true give verbose output}

\item{errors}{If True give errors as warning}

\item{dbhandle}{The database Handle, see #' \code{\link{PgObjectsInit}}}
}
\value{
the result of the SQL query as a data.frame or NULL in case of
a non SELECT query. If the query results in an error, NA is
returned.
this function is the workhorse of the psqrepo package, it runs
sql queries at the RDBMS and returns the result. This function
tries to catch any error and then returns NA. It is up to the user 
to test te result of the function.
}
\description{
Wrapper function to execute sql queries on a postgres database
}
