% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_storeKey.R
\name{pr_storeKey}
\alias{pr_storeKey}
\title{Store meta data key and value in repository}
\usage{
pr_storeKey(name, key, value, overwrite = FALSE)
}
\arguments{
\item{name}{Name of the object.}

\item{key}{Key.}

\item{value}{Value, must be of type character}

\item{overwrite}{If true, overwrite existing keys}
}
\value{
nothing

Meta data is stored as key, value pairs. This function associates a
key, value pair with the given object (objectname). Each object
can have an unlimited number of keys and it is up to the user to
organise the keys and possible values,

Values are always stored as character strings, so numeric values
must by changed into strings before they can be stored.

If a key allready exists then this function results in an error,
unless overwrite=TRUE
}
\description{
Meta data is stored as key,value pairs in the repository, This
function stores a key, value pair for a specific object
}
