% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_storeObj.R
\name{pr_storeObj}
\alias{pr_storeObj}
\title{Store an regular object into the repository}
\usage{
pr_storeObj(object, name, kv = NULL, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{object}{object te store in repository}

\item{name}{store object using thisname}

\item{kv}{meta data list with keyword - value pairs}

\item{overwrite}{overwrite object if exists}

\item{verbose}{give some verbose output}
}
\value{
NULL if succesfull

Any object existing in memory can be stored as object into the
psqlrepo repository. So you can store variables, data.frames,
functions etc. Some objects, like connections, can be stored but
this is less useful. Object which store data in external files,
like rasters or datatable objects, can't be stored into the
repository. For storing such objects the pr_storeBlob command can
be used. The name of the object can be any valid string.

TODO: explain something about kv

If an object allready exists and overwrite is FALSE this function
will return an error. If overwrite is TRUE then first it is checked
if the object exists and is equal to the object which will be
written, if so, the object is not writen to the database. If the
object to be written differs from the object in the database, the
object in the database is deleted and the new object is written.
}
\description{
This functions stores an regular object into the repository. Such
object can be any R object from memory
}
