% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_tableExists.R
\name{pr_tableExists}
\alias{pr_tableExists}
\title{Check if database table exists}
\usage{
pr_tableExists(table)
}
\arguments{
\item{table}{name of table}
}
\value{
True if table exists, false otherwise

This function checks if a table exists within the current psql
schema. If so, it returns True. If the table does not exists, it
returns false.

Existence of the table is tested using a SELECT query on that table
}
\description{
Check if database table exists
}
