% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_init.R
\name{pr_testConnection}
\alias{pr_testConnection}
\title{Check database connection}
\usage{
pr_testConnection(verbose = FALSE)
}
\value{
True, if connection exists, False otherwise

This function is used from other functions of this package. It
checks if a database connection exists and prints some information
about the connection.
}
\description{
This function checks if a database connection exists
}
